Add Authentication

+++ src/app/main/controller/board_controller.py
@@ -1,6 +1,7 @@
 from flask_restx import Resource
 from ..util.dto import BoardDto
 from ..service import board_service
+from ..service.auth_helper import Auth
 from flask import request
 
 api = BoardDto.api
@@ -12,8 +13,11 @@ class Board(Resource):
     def post(self):
         """create board"""
         try:
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
             data = request.json
-            data["user_id"] = 101
+            data["user_id"] = resp
             return board_service.add_board(data=data)
         except Exception as e:
             return {'status': 500,
@@ -25,7 +29,10 @@ class Board(Resource):
     def get(self, board_id):
         """get single board details"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
             return board_service.get_board(board_id=int(board_id), data=data)
         except Exception as e:
             return {'status': 500,
@@ -35,7 +42,10 @@ class Board(Resource):
     def delete(self, board_id):
         """get single board details"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
             return board_service.delete_board(board_id=int(board_id), data=data)
         except Exception as e:
             return {'status': 500,
@@ -48,9 +58,14 @@ class Board(Resource):
     def get(self):
         """List all boards"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
+            print("data",data)
             return board_service.get_all_boards(data)
         except Exception as e:
+            print(e)
             return {'status': 500,
                     'message': "Internal application error."}, 500
 
@@ -64,10 +79,14 @@ class Board(Resource):
     def post(self, board_id):
         """get single board details"""
         try:
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
             data = request.json
-            data["user_id"] = 101
+            data["user_id"] = resp
             return board_service.add_task_for_board(board_id=int(board_id), data=data)
         except Exception as e:
+            print(e)
             return {'status': 500,
                     'message': "Internal application error."}, 500
 
@@ -75,7 +94,10 @@ class Board(Resource):
     def get(self, board_id):
         """get board task based on filter"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
             data["query"] = request.args.get('query')
             data["status"] = request.args.get('status')
             data["priority"] = request.args.get('priority')
@@ -93,8 +115,11 @@ class Board(Resource):
     def post(self, board_id):
         """search for tasks"""
         try:
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
             data = request.json
-            data["user_id"] = 101
+            data["user_id"] = resp
             return board_service.search_tasks(board_id=int(board_id), data=data)
         except Exception as e:
             return {'status': 500,
@@ -106,7 +131,10 @@ class Board(Resource):
     def get(self, board_id, task_id):
         """get single board details"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
             return board_service.get_task(board_id=int(board_id), task_id=task_id, data=data)
         except Exception as e:
             return {'status': 500,
@@ -116,8 +144,11 @@ class Board(Resource):
     def put(self, board_id, task_id):
         """update single task details"""
         try:
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
             data = request.json
-            data["user_id"] = 101
+            data["user_id"] = resp
             return board_service.update_task_for_board(board_id=int(board_id), task_id=task_id, data=data)
         except Exception as e:
             return {'status': 500,
@@ -127,7 +158,10 @@ class Board(Resource):
     def delete(self, board_id, task_id):
         """archive single task details"""
         try:
-            data = {"user_id": 101}
+            resp = Auth.get_user_id(request.headers)
+            if not isinstance(resp, int):
+                return resp
+            data = {"user_id": resp}
             return board_service.archive_task_for_board(board_id=int(board_id), task_id=task_id, data=data)
         except Exception as e:
             return {'status': 500,
